requirements.txt
asgiref==3.8.1
boto3==1.34.100
botocore==1.34.100
cffi==1.16.0
cryptography==42.0.7
dj-database-url==2.1.0
Django==5.0.6
django-environ==0.11.2
django-storages==1.14.3
djangorestframework==3.15.1
jmespath==1.0.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
